+++ src/builtin.ml
@@ -1198,6 +1198,27 @@ set,
     (fun m _ ~depth -> !: (Set.elements m))),
   DocAbove);
 
+  MLCode(Pred(name^".choose",
+    In(set,"M",
+    Out(alpha,"X",
+    Easy "X is an element of M")),
+    (fun m _ ~depth -> !: (try Set.choose m with Not_found -> raise No_clause))),
+  DocAbove);
+
+  MLCode(Pred(name^".min",
+    In(set,"M",
+    Out(alpha,"X",
+    Easy "X is the smallest element of M")),
+    (fun m _ ~depth -> !: (try Set.min_elt m with Not_found -> raise No_clause))),
+  DocAbove);
+
+  MLCode(Pred(name^".max",
+    In(set,"M",
+    Out(alpha,"X",
+    Easy "X is the bigger of M")),
+    (fun m _ ~depth -> !: (try Set.max_elt m with Not_found -> raise No_clause))),
+  DocAbove);
+
   MLCode(Pred(name^".cardinal",
     In(set,"M",
     Out(int,"N",
@@ -1229,6 +1250,35 @@ set,
       
       state, !: m, gls
     )),
+  DocAbove);
+
+  MLCode(Pred(name^".fold",
+    In(set,"M",
+    In(poly "A","Acc",
+    In(HOAdaptors.pred2a alpha "A","F",
+    Out(poly "A","Acc1",
+    FullHO(ContextualConversion.unit_ctx, "fold M w.r.t. the predicate F"))))),
+    (fun m a f _ ~once ~depth _ _ state ->
+
+      let state, a, gls = HOAdaptors.fold1 ~once ~depth ~fold:Set.fold f m a state in
+      
+      state, !: a, gls
+    )),
+  DocAbove);
+
+
+  MLCode(Pred(name^".partition",
+  In(set,"M",
+  In(HOAdaptors.pred1 alpha,"F",
+  Out(set,"M1",
+  Out(set,"M2",
+  FullHO(ContextualConversion.unit_ctx, "Partitions M w.r.t. the predicate F, M1 is where F holds"))))),
+  (fun m f _ _ ~once ~depth _ _ state ->
+
+    let state, (m1, m2), gls = HOAdaptors.filter1 ~once ~depth ~filter:Set.partition f m state in
+    
+    state, !: m1 +! m2, gls
+  )),
   DocAbove);
 
 
@@ -1332,6 +1382,21 @@ let open BuiltIn in let open BuiltInData in
       let state, m, gls = HOAdaptors.map2 ~once ~depth ~map:Map.mapi f m state in
       
       state, !: m, gls
+    )),
+  DocAbove);
+
+
+  MLCode(Pred(name^".fold",
+    In(map "A","M",
+    In(poly "C","Acc",
+    In(HOAdaptors.pred3a alpha closed_A "C","F",
+    Out(poly "C","Acc1",
+    FullHO(ContextualConversion.unit_ctx, "fold M w.r.t. the predicate F"))))),
+    (fun m a f _ ~once ~depth _ _ state ->
+
+      let state, a, gls = HOAdaptors.fold2 ~once ~depth ~fold:Map.fold f m a state in
+      
+      state, !: a, gls
     )),
   DocAbove);